requirements.txt
vosk
sentence-transformers
faiss-cpu
pandas
numpy
torch
transformers
 